Moves test runner to test/
