+++ index.test.js
@@ -1,32 +1,5 @@
-const postcss = require('postcss');
-const tidyColumns = require('.');
-const fs = require('fs');
-const path = require('path');
 const fixtures = require('./test/fixtures/_fixtures.json');
-
-/**
- * Basic plugin test.
- * Run the plugin and return the output.
- */
-const run = (input, output, opts, plugin = tidyColumns) => (
-  postcss([plugin(opts)])
-    .process(input, { from: undefined })
-    .then((result) => {
-      expect(result.css).toEqual(output);
-      expect(result.warnings().length).toBe(0);
-    })
-);
-
-/**
- * Read file utility for shorter line-lengths.
- *
- * @param {String} filename The name of the file to read.
- *
- * @return {String}
- */
-function readFile(filename) {
-  return fs.readFileSync(path.join(__dirname, filename), 'utf8');
-}
+const { run, readFile } = require('./test');
 
 /**
  * Test fixtures
@@ -58,8 +31,3 @@ test(
     {},
   ),
 );
-
-module.exports = {
-  run,
-  readFile,
-};
