+++ lib/tidy-function.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-const { run } = require('../index.test.js');
+const { run } = require('../test');
 const { typical } = require('../test/sharedConfigs');
 
 describe('Test `tidy-offset` function replacement', () => {
