+++ lib/utils/cleanClone.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-const { run } = require('../../index.test.js');
+const { run } = require('../../test');
 const postcss = require('postcss');
 const cleanClone = require('./cleanClone');
